1581 clean up core components, update Dagger in Versions
