+++ core/src/main/java/org/kiwix/kiwixmobile/core/history/HistoryActivity.java
@@ -39,13 +39,11 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import javax.inject.Inject;
-import org.jetbrains.annotations.NotNull;
 import org.kiwix.kiwixmobile.core.CoreApp;
 import org.kiwix.kiwixmobile.core.Intents;
 import org.kiwix.kiwixmobile.core.R;
 import org.kiwix.kiwixmobile.core.R2;
 import org.kiwix.kiwixmobile.core.base.BaseActivity;
-import org.kiwix.kiwixmobile.core.di.components.HistoryActivityComponent;
 import org.kiwix.kiwixmobile.core.extensions.ImageViewExtensionsKt;
 import org.kiwix.kiwixmobile.core.main.CoreMainActivity;
 import org.kiwix.kiwixmobile.core.reader.ZimReaderContainer;
@@ -63,7 +61,7 @@ public class HistoryActivity extends BaseActivity implements HistoryContract.Vie
 
   @BindView(R2.id.toolbar)
   Toolbar toolbar;
-  @Inject
+  //@Inject
   HistoryContract.Presenter presenter;
   @Inject
   ZimReaderContainer zimReaderContainer;
@@ -126,13 +124,10 @@ public class HistoryActivity extends BaseActivity implements HistoryContract.Vie
       }
     }
   };
-  HistoryActivityComponent historyActivityComponent;
   @Override
   protected void onCreate(Bundle savedInstanceState) {
-
-    historyActivityComponent = CoreApp.getCoreComponent().historyActivityComponent().build();
-    historyActivityComponent.inject(this);
-
+    //Dagger injection
+    CoreApp.getCoreComponent().inject(this);
     super.onCreate(savedInstanceState);
     presenter.attachView(this);
     setContentView(R.layout.activity_history);
